Pull out make_coref_instance method
